Create LICENSE.md
Your selection highlighter is awesome, and I want to get it in use by lots of developers at my company.  Sadly our legal team always needs a EULA / license.md file for open-source code.  I've taken the liberty of drafting you a permissive BSD3 license, which should protect you as well as allow us to use it.  Feel free to switch it to MIT or Apache 2.0 if you prefer.